Modularise generated http client and types for economy codegen
