+++ packages/economy/sdk/http-client/item-definition-api/http-client.ts
@@ -9,99 +9,6 @@
  * ---------------------------------------------------------------
  */
 
-export interface DomainCondition {
-  comparison?: string;
-  expected?: string;
-  ref?: string;
-  type?: string;
-}
-
-export interface DomainInput {
-  conditions?: DomainCondition[];
-  id?: string;
-  name?: string;
-  type?: string;
-}
-
-export interface DomainOutput {
-  data?: object;
-  id?: string;
-  location?: string;
-  name?: string;
-  ref?: string;
-  type?: string;
-}
-
-export interface DomainRecipe {
-  created_at?: string;
-  description?: string;
-  game_id?: string;
-  id?: string;
-  inputs?: DomainInput[];
-  name?: string;
-  outputs?: DomainOutput[];
-  status?: string;
-  updated_at?: string;
-}
-
-export interface HttpCondition {
-  comparison?: string;
-  expected?: string;
-  ref?: string;
-  type?: string;
-}
-
-export interface HttpCreateRecipeInput {
-  conditions?: HttpCondition[];
-  name?: string;
-  /** @example "single_item" */
-  type?: 'single_item' | 'multiple_item' | 'single_currency' | 'multiple_currency';
-}
-
-export interface HttpCreateRecipeOutput {
-  data?: object;
-  location?: string;
-  name?: string;
-  ref?: string;
-  /** @example "item_definition" */
-  type?: 'item_definition' | 'input';
-}
-
-export interface HttpCreateRecipeRequest {
-  description?: string;
-  /** @example "shardbound" */
-  game_id: string;
-  inputs?: HttpCreateRecipeInput[];
-  name?: string;
-  outputs?: HttpCreateRecipeOutput[];
-}
-
-export interface HttpUpdateRecipeInput {
-  conditions?: HttpCondition[];
-  id?: string;
-  name?: string;
-  /** @example "single_item" */
-  type?: 'single_item' | 'multiple_item' | 'single_currency' | 'multiple_currency';
-}
-
-export interface HttpUpdateRecipeOutput {
-  data?: object;
-  id?: string;
-  location?: string;
-  name?: string;
-  ref?: string;
-  /** @example "item_definition" */
-  type?: 'item_definition' | 'input';
-}
-
-export interface HttpUpdateRecipeRequest {
-  description?: string;
-  inputs?: HttpUpdateRecipeInput[];
-  name?: string;
-  outputs?: HttpUpdateRecipeOutput[];
-  status?: string;
-}
-
 export type QueryParamsType = Record<string | number, any>;
 export type ResponseFormat = keyof Omit<Body, 'body' | 'bodyUsed'>;
 
@@ -148,7 +55,7 @@ export enum ContentType {
 }
 
 export class HttpClient<SecurityDataType = unknown> {
-  public baseUrl: string = '/recipe/v1';
+  public baseUrl: string = '/item-definition/v1';
   private securityData: SecurityDataType | null = null;
   private securityWorker?: ApiConfig<SecurityDataType>['securityWorker'];
   private abortControllers = new Map<CancelToken, AbortController>();
@@ -309,115 +216,5 @@ export class HttpClient<SecurityDataType = unknown> {
       if (!response.ok) throw data;
       return data;
     });
-  };
-}
-
-/**
- * @title Recipe API
- * @version 1.0
- * @baseUrl /recipe/v1
- * @contact
- *
- * Recipe API
- */
-export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
-  recipes = {
-    /**
-     * @description Get all recipes
-     *
-     * @tags root
-     * @name RecipesList
-     * @summary Get all recipes
-     * @request GET:/recipes
-     */
-    recipesList: (
-      query: {
-        /** Game ID */
-        game_id: string;
-        /**
-         * Map of conditions and values
-         * @example "condition.item_definition_id=XYZ"
-         */
-        filters?: string[];
-      },
-      params: RequestParams = {},
-    ) =>
-      this.request<DomainRecipe[], any>({
-        path: `/recipes`,
-        method: 'GET',
-        query: query,
-        type: ContentType.Json,
-        format: 'json',
-        ...params,
-      }),
-
-    /**
-     * @description Create recipe
-     *
-     * @tags root
-     * @name RecipesCreate
-     * @summary Create recipe
-     * @request POST:/recipes
-     */
-    recipesCreate: (request: HttpCreateRecipeRequest, params: RequestParams = {}) =>
-      this.request<DomainRecipe, any>({
-        path: `/recipes`,
-        method: 'POST',
-        body: request,
-        type: ContentType.Json,
-        format: 'json',
-        ...params,
-      }),
-
-    /**
-     * @description Get recipe by id
-     *
-     * @tags root
-     * @name RecipesDetail
-     * @summary Get recipe by id
-     * @request GET:/recipes/{id}
-     */
-    recipesDetail: (id: string, params: RequestParams = {}) =>
-      this.request<DomainRecipe, any>({
-        path: `/recipes/${id}`,
-        method: 'GET',
-        type: ContentType.Json,
-        format: 'json',
-        ...params,
-      }),
-
-    /**
-     * @description Update recipe
-     *
-     * @tags root
-     * @name RecipesUpdate
-     * @summary Update recipe
-     * @request PUT:/recipes/{id}
-     */
-    recipesUpdate: (id: string, request: HttpUpdateRecipeRequest, params: RequestParams = {}) =>
-      this.request<DomainRecipe, any>({
-        path: `/recipes/${id}`,
-        method: 'PUT',
-        body: request,
-        type: ContentType.Json,
-        format: 'json',
-        ...params,
-      }),
-
-    /**
-     * @description Delete recipe by id
-     *
-     * @tags root
-     * @name RecipesDelete
-     * @summary Delete recipe by id
-     * @request DELETE:/recipes/{id}
-     */
-    recipesDelete: (id: string, params: RequestParams = {}) =>
-      this.request<void, any>({
-        path: `/recipes/${id}`,
-        method: 'DELETE',
-        type: ContentType.Json,
-        ...params,
-      }),
   };
 }
