+++ packages/economy/sdk/package.json
@@ -48,10 +48,10 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "typecheck": "tsc --noEmit --jsx preserve",
-    "codegen:item-definition": "rm -rf ./http-client/item-definition-api.ts; swagger-typescript-api -p $npm_package_config_item_definition -o ./http-client -n item-definition-api.ts",
-    "codegen:crafting": "rm -rf ./http-client/crafting-api.ts; swagger-typescript-api -p $npm_package_config_crafting -o ./http-client -n crafting-api.ts",
-    "codegen:recipes": "rm -rf ./http-client/recipes-api.ts; swagger-typescript-api -p $npm_package_config_recipes -o ./http-client -n recipes-api.ts",
-    "codegen:inventory": "rm -rf ./http-client/inventory-api.ts; swagger-typescript-api -p $npm_package_config_inventory -o ./http-client -n inventory-api.ts",
+    "codegen:item-definition": "rm -rf ./http-client/item-definition-api; swagger-typescript-api -p $npm_package_config_item_definition -o ./http-client/item-definition-api --modular",
+    "codegen:crafting": "rm -rf ./http-client/crafting-api; swagger-typescript-api -p $npm_package_config_crafting -o ./http-client/crafting-api --modular",
+    "codegen:recipes": "rm -rf ./http-client/recipes-api; swagger-typescript-api -p $npm_package_config_recipes -o ./http-client/recipes-api --modular",
+    "codegen:inventory": "rm -rf ./http-client/inventory-api; swagger-typescript-api -p $npm_package_config_inventory -o ./http-client/inventory-api --modular",
     "codegen": "rm -rf ./http-client; npm run codegen:item-definition & npm run codegen:crafting & npm run codegen:recipes & npm run codegen:inventory"
   },
   "type": "module",
